Move lib and utils into api and update imports
