+++ api/contacts/index.ts
@@ -3,9 +3,9 @@ import path from "path";
 console.log("CWD:", process.cwd());
 console.log("Files at root:", fs.readdirSync(process.cwd()));
 console.log("Files in lib:", fs.existsSync("lib") ? fs.readdirSync("lib") : "lib missing");
-import getAirtableContext from "../../lib/airtableBase";
-import { getFieldMap } from "../../lib/resolveFieldMap";
-import { mapInternalToAirtable } from "../../lib/mapRecordFields";
+import getAirtableContext from "../lib/airtableBase";
+import { getFieldMap } from "../lib/resolveFieldMap";
+import { mapInternalToAirtable } from "../lib/mapRecordFields";
 import { FieldSet, Record as AirtableRecord } from "airtable";
 
 const apiContactsHandler = async (req: any, res: any) => {
