+++ api/snapshots/index.ts
@@ -1,6 +1,6 @@
-import getAirtableContext from "../../lib/airtableBase";
-import { getFieldMap } from "../../lib/resolveFieldMap";
-import { mapInternalToAirtable } from "../../lib/mapRecordFields";
+import getAirtableContext from "../lib/airtableBase";
+import { getFieldMap } from "../lib/resolveFieldMap";
+import { mapInternalToAirtable } from "../lib/mapRecordFields";
 import { FieldSet, Record as AirtableRecord } from "airtable";
